feat: split app.rs into multiple files
